Added observer methods to use the logged in user.
Right now, this is just assuming there is a user with the ID of 1. This will be further fleshed out in the upcoming seeders. I understand that this may be jumping the gun in terms of priority, but I want to have all that in place, as most everything else will revolve around the users.

Once this is done, I'll remove the Auth::loginUsingID() call in AppServiceProvider.php.